Update as of 20150915

+++ BaseProject/src/project/dto/EventModelDto.java
@@ -16,7 +16,7 @@ private static final long serialVersionUID = 1L;
     /**
      * Event ID container
      */
-    private int eventID;
+    private long eventID;
     /**
      *  Event Name container.
      */
@@ -29,13 +29,13 @@ private static final long serialVersionUID = 1L;
     /**
      * @return the event ID
      */
-    public int getEventID(){
+    public long getEventID(){
         return eventID;
     }
     /**
      * @param id the eventID to set
      */
-    public void setEventID(int id){
+    public void setEventID(long id){
         this.eventID = id;       
     }
     /**
